==> 00--required_data.Rmd <==
---
output: html_document
bibliography: "bibtexlib.bib"
---

```{r setup, include=FALSE}
source("style.R")
```

# Datasets used in this workshop

In this workshop, we will be using data from @wagner2016host, a study on the effects of plant age, genotype, and environment on the bacterial microbiome of [*Boechera stricta*](https://en.wikipedia.org/wiki/Boechera_stricta), a perennial herb in the mustard family.
@wagner2016host released their raw data with the article and it is available [here](http://datadryad.org/resource/doi:10.5061/dryad.g60r3) on [dryad](http://datadryad.org/).
This is a great example of how to share your raw data!

A copy of this data is included in this site and can be downloaded from links below.

## Sample metadata

A table with a sample in each row and info about the samples in each column, such as plant genotype and field site.
It is a 170Kb `r gloss$add('tab-delimited text file')`.

<a href="data/SMD.txt" download="data/SMD.txt">SMD.txt</a>

## The OTU abundance table

This has the number of reads associated with each `r gloss$add('Operational Taxonomic Units (OTUs)', shown = 'OTU')` in each sample.
It is a 6Mb compressed tab-delimited text file.

<a href="data/otuTable97.txt.bz2" download="data/otuTable97.txt.bz2">otuTable97.txt.bz2</a>

## The OTU taxonomy file

This has the `r gloss$add('taxonomic classifications')` for each OTU.
It is a 6Mb tab-delimited text file.

<a href="data/taxAssignments97.txt" download="data/taxAssignments97.txt">taxAssignments97.txt</a>

## References
